Udpate tests related to link definitions
